// src/app/shared/models/note.model.ts
export class note{
    id : number;
    title: string;
    content: string;
    createddate: Date;
    createdby: number;
    updateddate: Date;
    updatedby: number;
}

export class Note{
    id : number;
    title: string;
    content: string;
    createddate: Date;
    createdby: number;
    updateddate: Date;
    updatedby: number;
}